CRUD on Goals resource with logged in User Authorization Completed

+++ backend/controllers/goalController.js
@@ -1,13 +1,17 @@
 const asyncHandler = require('express-async-handler');
 const Goal = require('../models/goalModel');
+const User = require('../models/userModel');
 
 
 // @desc    Get goals
 // @route   GET /api/goals
 // @access  Private
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find()
-  res.status(200).json(goals)
+  // In the below database find we can use
+  // req.user because we have used the protect middleware on getGoals
+  const goals = await Goal.find({ user: req.user.id });
+
+  res.status(200).json(goals);
 })
 
 // @desc    Create goals
@@ -20,7 +24,8 @@ const setGoal = asyncHandler(async (req, res) => {
   }
 
   const goal = await Goal.create({
-    text: req.body.text
+    text: req.body.text,
+    user: req.user.id
   })
 
   res.status(201).json(goal);
@@ -35,6 +40,22 @@ const updateGoal = asyncHandler(async (req, res) => {
     res.status(400)
     throw new Error('Goal not Found')
   }
+
+  // Find the user who is logged in by using the protect middleware's req.user
+  const user = await User.findById(req.user.id);
+
+  // Check for user
+  if (!user) {
+    res.status(401);
+    throw new Error('User not found');
+  }
+
+  // Make sure the logged in user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error('User not authorized');
+  }
+
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new: true});
   res.status(200).json(updatedGoal);
 })
@@ -48,6 +69,22 @@ const deleteGoal = asyncHandler(async (req, res) => {
     res.status(400)
     throw new Error('Goal not Found')
   }
+
+  // Find the user who is logged in by using the protect middleware's req.user
+  const user = await User.findById(req.user.id);
+
+  // Check for user
+  if (!user) {
+    res.status(401);
+    throw new Error('User not found');
+  }
+
+  // Make sure the logged in user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error('User not authorized');
+  }
+
 
   await Goal.findByIdAndDelete(req.params.id);
   res.status(200).json({message: `id:  ${req.params.id}`})
